Consolidate even more rendering code to the rendering system
The monster ai system and Game class both had code that modified room.
This used to be because the room component's @room used to be mutated.

Now, @room is deep duped so that the rendering system can modify it
without those changes needing to be handled elsewhere.

At the end of the process_one_game_tick method the duped room is thrown
away.