Kind of awkward hack to fix PlayerInputs cascading and causing each other to trigger
The fix is that instead of the PlayerInput callback returning early if
some conditions aren't met, there is now a separate method to call that
returns whether the conditions are met.

This is important because it lets the PlayerInputSystem filter all of
the PlayerInput components to only those that have their conditions met
(using the current state of the game), _then_ runs those. When going
from level_1 to level_2, only the first PlayerInput gets run, instead of
both.

The code is pretty meh, but should be fine for now. Once there's lots of
these PlayerInput class methods, going to need better solution.